% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaTest.R
\name{anova_test}
\alias{anova_test}
\title{A pipe friendly wrapper for one way ANOVA}
\usage{
anova_test(data, indep_var)
}
\arguments{
\item{data}{The data frame containing variables to be analyzed}

\item{indep_var}{The independent or predictor variable in the data}
}
\value{
dataframe containing the ANOVA result
}
\description{
It carries out one way ANOVA and is pipe friendly
}
