% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binder.R
\name{binder_1}
\alias{binder_1}
\title{Bind the spliited final results into a dataframe}
\usage{
binder_1(ls_summary, grouping_vars)
}
\arguments{
\item{ls_summary}{splitted final results from table_summary function in
meanSeparator class}

\item{grouping_vars}{grouping vars specified in meanSeparator class}
}
\value{
dataframe
}
\description{
It combines the list of results splitted by grouping vars into
a single dataframe
}
