% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binder.R
\name{binder_2}
\alias{binder_2}
\title{Bind the result of post-hoc and aov into a dataframe}
\usage{
binder_2(post_hoc_tbl, aov_tbl)
}
\arguments{
\item{post_hoc_tbl}{final results from meanSeparator}

\item{aov_tbl}{aov results}
}
\value{
dataframe
}
\description{
It combines the list of results splitted by grouping vars into
a single dataframe
}
