% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{custom_split}
\alias{custom_split}
\title{A split function adapted for multiple split groups}
\usage{
custom_split(data, x)
}
\arguments{
\item{data}{data to be splited}

\item{x}{vector or list of data that will be used to split the data}
}
\value{
list of splited data
}
\description{
A split function adapted for multiple split groups
}
