% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaTest_2.R
\name{fastanova_test_2}
\alias{fastanova_test_2}
\title{A more complex anova function, thats carries out one-way
, two-way and mixed anova}
\usage{
fastanova_test_2(
  data,
  wid = NULL,
  within = NULL,
  between = NULL,
  term = "all",
  btw_terms_sep = "+",
  within_terms_sep = "*",
  transform_func = NULL
)
}
\arguments{
\item{data}{dataframe containing data to analyzed}

\item{wid}{the subject identifiers in the datasets}

\item{within}{variables used for within effect of the analysis}

\item{between}{variables used for between effect of the analysis}

\item{term}{the type of term to extract which is one of `all` - gets all effect except `residuals`, # nolint
`marginal`- gets only marginal effect and `interaction` - gets only interaction effects # nolint}

\item{btw_terms_sep}{between terms formula seprator with could be `+`, `*` etc # nolint}

\item{within_terms_sep}{=within terms formula seprator with could be `+`, `*` etc # nolint}

\item{transform_func}{a callback that can be to transorm each dependent var in the dataset # nolint}
}
\value{
a matrix containing p values
}
\description{
A more complex anova function, thats carries out one-way
, two-way and mixed anova
}
