% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummaryStats.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Generate summary statistic with ANOVA for a data set}
\usage{
generate_summary_stats(
  data,
  indep_var,
  deviation_type = "s.e",
  grouping_vars = NA,
  factor_vars = NA
)
}
\arguments{
\item{data}{The data frame containing variables to be analyzed}

\item{indep_var}{The independent or predictor variable in the data}

\item{deviation_type}{The type of degree of spread - `s.e` or `sd` default to `s.e`}

\item{grouping_vars}{The grouping variables in the data if there are any - Optional}

\item{factor_vars}{The factor variables in the data - Optional when `grouping var` argument is specified}
}
\value{
a List of summary tables for all groups
}
\description{
It generate summary statistics and attaches ANOVA p values to the table and also ranks the data set
}
