% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSeperator.R
\docType{class}
\name{MeanSeperator-class}
\alias{MeanSeperator-class}
\alias{MeanSeperator}
\title{Mean seperator using tukey HSD}
\value{
a List
Display summary statistics with mean of separation

display human readable format of the separated mean

a Data frame
}
\description{
Is contains function that separate means and display them in a data frame
}

