% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_repititions}
\alias{remove_repititions}
\title{Convert repetitions into elipsis}
\usage{
remove_repititions(df_, vars)
}
\arguments{
\item{df_}{dataframe}

\item{vars}{columns in the dataframe to focus on}
}
\value{
dataframe
}
\description{
Convert repetitions into elipsis
}
