% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSeperator.R
\name{Separator}
\alias{Separator}
\title{Mean seperator using tukey HSD}
\description{
Is contains function that separate means and display them in a data frame
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The data frame containing variables to be analyzed}

\item{\code{x}}{The independent or predictor variable in the data}

\item{\code{factor_vars}}{The factor variables in the data - Optional when `grouping var` argument is specified # nolint}

\item{\code{grouping_vars}}{The grouping variables in the data if there are any - Optional # nolint}

\item{\code{deviation_type}}{The type of degree of spread - `s.e` or `sd` default to `s.e` # nolint}

\item{\code{code_seperator}}{The internal code separator defaults to `@` - Optional}

\item{\code{letter.name}}{The column name for the separation of mean letters}

\item{\code{decreasing}}{The order at which the alphabet are ranked - defaults to `FALSE` # nolint}

\item{\code{format}}{The representation of the summary data in the dataframe - defaults to `plain` - options are `plain`, `md`, and `html` # nolint}

\item{\code{include}}{a vector containing ANOVA statistical result to add to the final data frame - `f-value` or/and `p-value` defaults to only `p-value` # nolint}

\item{\code{transform_func}}{a callback that can be to transorm each dependent var in the dataset # nolint}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Separator-new}{\code{Separator$new()}}
\item \href{#method-Separator-separate}{\code{Separator$separate()}}
\item \href{#method-Separator-compute_ANOVA}{\code{Separator$compute_ANOVA()}}
\item \href{#method-Separator-display_table}{\code{Separator$display_table()}}
\item \href{#method-Separator-table_summary}{\code{Separator$table_summary()}}
\item \href{#method-Separator-clone}{\code{Separator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$new(
  data,
  x,
  factor_vars = NULL,
  grouping_vars = NULL,
  deviation_type = "s.e",
  format = "plain",
  include = "p-value",
  decreasing = FALSE,
  transform_func = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing variables to be analyzed}

\item{\code{x}}{The independent or predictor variable in the data}

\item{\code{factor_vars}}{The factor variables in the data - Optional when `grouping var` argument is specified # nolint}

\item{\code{grouping_vars}}{The grouping variables in the data if there are any - Optional # nolint}

\item{\code{deviation_type}}{The type of degree of spread - `s.e` or `sd` default to `s.e` # nolint}

\item{\code{format}}{The representation of the summary data in the dataframe - defaults to `plain` - options are `plain`, # nolint}

\item{\code{include}}{a vector containing ANOVA statistical result to add to the final data frame - `f-value` or/and `p-value` defaults to only `p-value` # nolint}

\item{\code{decreasing}}{The order at which the alphabet are ranked - defaults to `FALSE` # nolint}

\item{\code{transform_func}}{a callback that can be to transorm each dependent var in the dataset # nolint
separates the dataframe

It separates and return a list of the results in there respective groups}

\item{\code{console_view}}{print as plain text if set to `TRUE` or markdown if set to `FALSE` # nolint}

\item{\code{code_seperator}}{The internal code separator defaults to `@` - Optional # nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-separate"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-separate}{}}}
\subsection{Method \code{separate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$separate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list containing the groups of data set result
Compute the ANOVA for the data

it compute the anova and returns a dataframe containing
groups and p-value's
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-compute_ANOVA"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-compute_ANOVA}{}}}
\subsection{Method \code{compute_ANOVA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$compute_ANOVA()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataframe
displays a human readable table

It display human readable dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-display_table"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-display_table}{}}}
\subsection{Method \code{display_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$display_table(order_by = "x", rep_rm = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order_by}}{Option to either order the results using independent variable (x) or the grouping vars # nolint}

\item{\code{rep_rm}}{logical argument indicating to remove repitions from grouping variables # nolint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe
Create a table summary for all the groups

It create a summary of all the possible groups if there are any
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-table_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-table_summary}{}}}
\subsection{Method \code{table_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$table_summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list | dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Separator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Separator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Separator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
