% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyHSD.R
\name{tukey.HSD}
\alias{tukey.HSD}
\title{it is a pipe friendly version of the base R TukeyHSD function}
\usage{
tukey.HSD(data, formula)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found. If missing, the variables are searched for in the standard way.}

\item{formula}{A formula specifying the model}
}
\value{
a dataframe
}
\description{
it is a pipe friendly version of the base R TukeyHSD function
}
