% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vec.na.rm}
\alias{vec.na.rm}
\title{Helper function to help remove na's and null in a vector}
\usage{
vec.na.rm(vec)
}
\arguments{
\item{vec}{vector to be cleaned}
}
\value{
vector
}
\description{
Helper function to help remove na's and null in a vector
}
