% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{get_summary}
\alias{get_summary}
\title{Format descriptive stats summary}
\usage{
get_summary(x, spread = "s.e")
}
\arguments{
\item{x}{numeric vector to be summarized}

\item{spread}{deviation type}
}
\value{
character vector
}
\description{
It finds the mean and standard error of a numeric vector
and formats then using  +/-.
}
