% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSeperator.R
\docType{class}
\name{Separator-class}
\alias{Separator-class}
\alias{Separator}
\title{Mean seperator using tukey HSD}
\arguments{
\item{data}{The data frame containing variables to be analyzed}

\item{indep_var}{The independent or predictor variable in the data}

\item{factor_var}{The factor variables in the data - Optional when `grouping var` argument is specified}

\item{grouping_var}{The grouping variables in the data if there are any - Optional}

\item{deviation_type}{The type of degree of spread - `s.e` or `sd` default to `s.e`}

\item{code_seperator}{The internal code separator defaults to `@` - Optional}
}
\description{
Is contains function that separate means and display them in a data frame
}

